pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vm_tool"
version = "1.1.0"
description = "A Comprehensive Tool for Setting Up Virtual Machines."
readme = "README.md"
authors = [
    "Sunny Sinha <thesunnysinha@gmail.com>",
]
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
packages = [{include = "vm_tool"}]
include = ["vm_tool/vm_setup/**/*"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
ansible = ">=2.15,<11.0"
ansible-runner = ">=2.3,<3.0"
paramiko = ">=2.11,<4.0"
pydantic = ">=2.0,<3.0"
pyyaml = ">=6.0,<7.0"
tenacity = ">=8.2,<10.0"
rich = ">=13.0,<15.0"
python-dotenv = ">=1.0,<2.0"

# Optional cloud/k8s providers
boto3 = {version = ">=1.26,<2.0", optional = true}
google-cloud-compute = {version = ">=1.14,<2.0", optional = true}
azure-mgmt-compute = {version = ">=30.0,<32.0", optional = true}
azure-identity = {version = ">=1.12,<2.0", optional = true}
fabric = {version = ">=3.2,<4.0", optional = true}
requests = {version = ">=2.28,<3.0", optional = true}

[tool.poetry.extras]
aws = ["boto3"]
gcp = ["google-cloud-compute"]
azure = ["azure-mgmt-compute", "azure-identity"]
alerting = ["requests"]
ssh = ["fabric"]
all = ["boto3", "google-cloud-compute", "azure-mgmt-compute", "azure-identity", "requests", "fabric"]

[tool.poetry.group.dev.dependencies]
pytest = ">=6.0.0"
check-manifest = "*"
twine = "*"
wheel = "*"
bump-my-version = "*"
bandit = "*"
pre-commit = "*"
mkdocs-material = "*"
mkdocstrings = {extras = ["python"], version = "*"}
black = "*"
isort = "*"
flake8 = "*"

[tool.poetry.scripts]
vm_tool = "vm_tool.cli:main"

[tool.poetry.urls]
Documentation = "https://vm-tool.sunnysinha.online/"
Source = "https://github.com/thesunnysinha/vm_tool"
Tracker = "https://github.com/thesunnysinha/vm_tool/issues"

[tool.semantic_release]
version_variable = [
    "pyproject.toml:version"
]
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "poetry build"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.bumpversion]
current_version = "1.1.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
